snapuserd: Remove dm-user specific code from ReadWorker.
This uses the new IBlockServer abstraction layer instead.

Bug: 288273605
Test: snapuserd_test

+++ fs_mgr/libsnapshot/snapuserd/include/snapuserd/block_server.h
@@ -38,7 +38,7 @@ class IBlockServer {
         //
         // If false is returned, an error will be automatically reported unless
         // SendError was called.
-        virtual bool RequestSectors(uint64_t sector, size_t size) = 0;
+        virtual bool RequestSectors(uint64_t sector, uint64_t size) = 0;
     };
 
     virtual ~IBlockServer() {}
